refactoring to libraries
