+++ 3DStreamingToolKit.sln
@@ -67,6 +67,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DXUTOpt", "Libraries\DXUT\O
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Authentication", "Libraries\Authentication\Authentication.vcxproj", "{0BF47A53-06E2-4102-B789-C36F46D95DC7}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UserInterface", "Libraries\UserInterface\UserInterface.vcxproj", "{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		Plugins\UnityClientPlugin\MediaEngineUWP\Shared\Shared.vcxitems*{4a859119-6730-4612-987f-dabf98f213ed}*SharedItemsImports = 4
@@ -357,6 +359,18 @@ Global
 		{0BF47A53-06E2-4102-B789-C36F46D95DC7}.Release|x64.Build.0 = Release|x64
 		{0BF47A53-06E2-4102-B789-C36F46D95DC7}.Release|x86.ActiveCfg = Release|Win32
 		{0BF47A53-06E2-4102-B789-C36F46D95DC7}.Release|x86.Build.0 = Release|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Debug|x64.ActiveCfg = Debug|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Debug|x64.Build.0 = Debug|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Debug|x86.ActiveCfg = Debug|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Debug|x86.Build.0 = Debug|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Profile|x64.ActiveCfg = Release|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Profile|x64.Build.0 = Release|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Profile|x86.ActiveCfg = Release|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Profile|x86.Build.0 = Release|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Release|x64.ActiveCfg = Release|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Release|x64.Build.0 = Release|x64
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Release|x86.ActiveCfg = Release|Win32
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -389,5 +403,6 @@ Global
 		{85344B7F-5AA0-4E12-A065-D1333D11F6CA} = {C1D9AA9A-9247-44AB-B59A-DEDA3DAD5C55}
 		{61B333C2-C4F7-4CC1-A9BF-83F6D95588EB} = {C1D9AA9A-9247-44AB-B59A-DEDA3DAD5C55}
 		{0BF47A53-06E2-4102-B789-C36F46D95DC7} = {C1D9AA9A-9247-44AB-B59A-DEDA3DAD5C55}
+		{95EB9544-2E62-4F54-8567-D55CD1DFD8FE} = {C1D9AA9A-9247-44AB-B59A-DEDA3DAD5C55}
 	EndGlobalSection
 EndGlobal
