+++ Samples/Client/DirectxWin32/StreamingDirectxClient.vcxproj
@@ -200,7 +200,6 @@
     <ClCompile Include="src\conductor.cpp" />
     <ClCompile Include="src\data_channel_handler.cpp" />
     <ClCompile Include="src\defaults.cpp" />
-    <ClCompile Include="src\default_main_window.cpp" />
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\win32_data_channel_handler.cpp" />
   </ItemGroup>
@@ -209,11 +208,9 @@
     <ClInclude Include="inc\conductor.h" />
     <ClInclude Include="inc\data_channel_handler.h" />
     <ClInclude Include="inc\defaults.h" />
-    <ClInclude Include="inc\default_main_window.h" />
     <ClInclude Include="inc\defs.h" />
     <ClInclude Include="inc\flagdefs.h" />
     <ClInclude Include="inc\macros.h" />
-    <ClInclude Include="inc\main_window.h" />
     <ClInclude Include="inc\win32_data_channel_handler.h" />
     <ClInclude Include="inc\webrtc.h" />
     <ClInclude Include="pch.h" />
@@ -226,5 +223,6 @@
   </ItemGroup>
   <Import Project="$(MSBuildThisFileDirectory)..\..\..\Libraries\SignalingClient\exports.props" />
   <Import Project="$(MSBuildThisFileDirectory)..\..\..\Libraries\Authentication\exports.props" />
+  <Import Project="$(MSBuildThisFileDirectory)..\..\..\Libraries\UserInterface\exports.props" />
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
 </Project>
