+++ Samples/Server/Directx-SpinningCube/App.cpp
@@ -6,6 +6,8 @@
 
 #include "DeviceResources.h"
 #include "CubeRenderer.h"
+
+#include "server_main_window.h"
 #include "server_authentication_provider.h"
 #include "turn_credential_provider.h"
 
@@ -149,9 +151,9 @@ int InitWebRTC(char* server, int port, int heartbeat,
 	rtc::ThreadManager::Instance()->SetCurrentThread(&w32_thread);
 
 #ifdef NO_UI
-	DefaultMainWindow wnd(server, port, true, true, true);
+	ServerMainWindow wnd(server, port, true, true, true);
 #else // NO_UI
-	DefaultMainWindow wnd(server, port, FLAG_autoconnect, FLAG_autocall, false, 1280, 720);
+	ServerMainWindow wnd(server, port, FLAG_autoconnect, FLAG_autocall, false, 1280, 720);
 #endif // NO_UI
 
 	if (!wnd.Create())
@@ -188,6 +190,12 @@ int InitWebRTC(char* server, int port, int heartbeat,
 			if (data.successFlag)
 			{
 				client.SetAuthorizationHeader("Bearer " + data.accessToken);
+
+				// indicate to the user auth is complete (only if turn isn't in play)
+				if (turnProvider.get() == nullptr)
+				{
+					wnd.SetAuthCode(L"OK");
+				}
 			}
 		});
 
@@ -200,7 +208,6 @@ int InitWebRTC(char* server, int port, int heartbeat,
 		new rtc::RefCountedObject<Conductor>(
 			&client, &wnd, &FrameUpdate, &InputUpdate, g_videoHelper));
 
-
 	if (!turnCredentialUri.empty())
 	{
 		turnProvider.reset(new TurnCredentialProvider(turnCredentialUri));
@@ -210,6 +217,9 @@ int InitWebRTC(char* server, int port, int heartbeat,
 			if (creds.successFlag)
 			{
 				conductor->SetTurnCredentials(creds.username, creds.password);
+
+				// indicate to the user turn is done
+				wnd.SetAuthCode(L"OK");
 			}
 		});
 
@@ -222,11 +232,18 @@ int InitWebRTC(char* server, int port, int heartbeat,
 
 			// if we have auth, first get it
 			authProvider->Authenticate();
+
+			// indicate to the user we're authenticating
+			wnd.SetAuthUri(std::wstring(authInfo.authority.begin(), authInfo.authority.end()));
+			wnd.SetAuthCode(L"Loading");
 		}
 		else
 		{
 			// no auth, just get creds
 			turnProvider->RequestCredentials();
+
+			// indicate to the user we're turning
+			wnd.SetAuthCode(L"Loading");
 		}
 	}
